requirements.txt
openai
pandas
python-dotenv
supabase
numpy
requests
